refactor(progress-signals): extract shared component-drift walk from crap/maintainability detectors (refs #3984)
